cdk_workshop/requirements.txt
aws-cdk-lib==2.117.0
constructs>=10.0.0,<11.0.0
numpy
opencv-python-headless
Pillow==8.4.0
